services/ai-query-engine/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.27.0
pydantic>=2.7.4
pydantic-settings>=2.3.0
pymysql==1.1.0
# 2.9.11+ required for Python 3.13 wheels
psycopg2-binary==2.9.11
# 9.x required for Python 3.13 wheels
mysql-connector-python==9.1.0
pymongo==4.6.1
# 5.3.0+ required for Python 3.13 wheels
pyodbc==5.3.0
# >= so pip resolves a version compatible with Python 3.13 + numpy 2.x
langchain>=0.3.0
langchain-google-genai>=2.0.0
cryptography>=42.0.0
python-dotenv==1.0.0
PyYAML>=6.0.1
# numpy 2.x required for Python 3.13 (1.x has no cp313 wheel)
numpy>=2.0.0
